Introduced interfaces and dependency injection

+++ src/main/java/com/mahmoud/booking/CarBookingService.java
@@ -5,6 +5,7 @@ import com.mahmoud.car.CarService;
 import com.mahmoud.user.User;
 import com.mahmoud.user.UserService;
 
+import java.io.IOException;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -13,9 +14,15 @@ import java.util.UUID;
 
 public class CarBookingService {
 
-    private static final CarBookingDao carBookingDao = new CarBookingDao();
-    private static final UserService userService = new UserService();
-    private static final CarService carService = new CarService();
+    private final CarBookingDao carBookingDao;
+    private final UserService userService;
+    private final CarService carService;
+
+    public CarBookingService(CarBookingDao carBookingDao, CarService carService, UserService userService) {
+        this.carBookingDao = carBookingDao;
+        this.carService = carService;
+        this.userService = userService;
+    }
 
     public CarBooking bookCar(UUID userId, UUID carId, LocalDateTime startDate, LocalDateTime endDate) {
 
@@ -34,56 +41,89 @@ public class CarBookingService {
 
         BigDecimal price = new BigDecimal(numberOfDays).multiply(car.getRentalPricePerDay());
 
+
         CarBooking carBooking = new CarBooking(UUID.randomUUID(), user, car, startDate, endDate, price, BookingStatus.ACTIVE, LocalDateTime.now());
 
-        carBookingDao.saveBooking(carBooking);
+        try {
+            carBookingDao.saveBooking(carBooking);
+        } catch (IOException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
 
         return carBooking;
     }
 
     private boolean isCarAvailable(UUID carId, LocalDateTime startDate) {
 
-        var bookings = carBookingDao.getAllBookings();
+        try {
+            CarBooking[] bookings = carBookingDao.getAllBookings();
 
-        for (CarBooking carBooking : bookings) {
-            if (carBooking.getCar().getId().equals(carId)) {
-                if ((carBooking.getEndDate().isAfter(startDate)) ||
-                        ((carBooking.getEndDate().isEqual(startDate)) ||
-                                !carBooking.getStatus().equals(BookingStatus.CANCELLED))) {
-                    return false;
+            for (CarBooking carBooking : bookings) {
+                if (carBooking.getCar().getId().equals(carId)) {
+                    if ((carBooking.getEndDate().isAfter(startDate)) ||
+                            ((carBooking.getEndDate().isEqual(startDate)) ||
+                                    !carBooking.getStatus().equals(BookingStatus.CANCELLED))) {
+                        return false;
+                    }
                 }
             }
+
+            return true;
+
+        } catch (IOException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
         }
-        return true;
     }
 
     public boolean deleteBooking(UUID bookingId) {
-        return carBookingDao.deleteBooking(bookingId);
+        try {
+
+            CarBooking[] bookings = carBookingDao.getAllBookings();
+
+            for (CarBooking carBooking : bookings) {
+                if (carBooking.getId().equals(bookingId))
+                    return carBookingDao.deleteBooking(bookingId);
+            }
+
+            return false;
+        } catch (IOException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public CarBooking[] getUserBookingById(UUID userId) {
 
         var size = 0;
 
-        var bookings = carBookingDao.getAllBookings();
+        try {
 
-        for (CarBooking carBooking : bookings) {
-            if (carBooking.getUser().getId().equals(userId)) size++;
-        }
+            CarBooking[] bookings = carBookingDao.getAllBookings();
 
-        CarBooking[] userBookings = new CarBooking[size];
-
-        for (int i = 0; i < userBookings.length; i++) {
-            if (bookings[i].getUser().getId().equals(userId)) {
-                userBookings[i] = bookings[i];
+            for (CarBooking carBooking : bookings) {
+                if (carBooking.getUser().getId().equals(userId)) size++;
             }
-        }
 
-        return userBookings;
+            CarBooking[] userBookings = new CarBooking[size];
+
+            for (int i = 0; i < userBookings.length; i++) {
+                if (bookings[i].getUser().getId().equals(userId)) {
+                    userBookings[i] = bookings[i];
+                }
+            }
+
+            return userBookings;
+
+        } catch (IOException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public CarBooking[] getAllBookings() {
-        return carBookingDao.getAllBookings();
+        try {
+            return carBookingDao.getAllBookings();
+        } catch (IOException | ClassNotFoundException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     private boolean validateDates(LocalDateTime start, LocalDateTime end) {
